test: fix my_system_fd handling of closed stdout and align CI exclusions
* cmake/native_c_tests/test_helper.h (my_system_fd): Add explicit handling
  for `&-` file redirects to properly close STDIN, STDOUT, and STDERR. This
  fixes intermittent timeouts in `c_test_close_stdout` where the harness
  was mistakenly opening a file literally named `&-`.
* cmake/native_c_tests/test_ginstall.c (test_ginstall_basic): Fix format
  specifier warnings by explicitly casting literal 0 to `size_t` for
  `ASSERT_EQ_FMT` strlen validation.
* .github/workflows/cmake.yml: Apply the `ctest -E` unstable test exclusions
  documented in `PORTING_TO_CMAKE_AND_CTEST.md` (and utilized in the
  Dockerfiles) to the GitHub Actions workflow to resolve platform-specific
  false-positive CI failures.

+++ cmake/native_c_tests/test_ginstall.c
@@ -59,11 +59,11 @@ TEST test_ginstall_C() {
     // destination file exists
     char out[1024];
     exec_capture("ginstall -Cv -m0644 a b", out, sizeof(out));
-    ASSERT_EQ_FMT(0, strlen(out), "%zu"); // should be empty output
+    ASSERT_EQ_FMT((size_t)0, strlen(out), "%zu"); // should be empty output
     
     // destination file exists (long option)
     exec_capture("ginstall -v --compare -m0644 a b", out, sizeof(out));
-    ASSERT_EQ_FMT(0, strlen(out), "%zu");
+    ASSERT_EQ_FMT((size_t)0, strlen(out), "%zu");
     
     // option -C ignored if any non-permission mode should be set
     exec_capture("ginstall -Cv -m2755 a b", out, sizeof(out));
@@ -76,7 +76,7 @@ TEST test_ginstall_C() {
     
     // second run should be empty
     exec_capture("ginstall -Cv -m0644 a b", out, sizeof(out));
-    ASSERT_EQ_FMT(0, strlen(out), "%zu");
+    ASSERT_EQ_FMT((size_t)0, strlen(out), "%zu");
     
     PASS();
 }
